requirements/base.txt
Django==4.1.5
django-environ==0.9.0
uvicorn==0.20.0
djangorestframework==3.14.0
